Use modulus rather than comparing square roots
Previously, this line checked whether sum divided by its square root
equaled its square root.  Due to floating point precision issues, this
isn't always strictly true, but it's not the question we want to be asking.

Rather, we should be asking whether the square root of sum is a whole number.

To answer that question, we need to use fmod()